pyproject.toml
[tool.poetry]
name = "vr180-convert"
version = "0.6.2"
description = "Simple VR180 image converter"
authors = ["34j <34j.95a2p@simplelogin.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/34j/vr180-convert"
documentation = "https://vr180-convert.readthedocs.io"
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries",
]
packages = [
    { include = "vr180_convert", from = "src" },
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/34j/vr180-convert/issues"
"Changelog" = "https://github.com/34j/vr180-convert/blob/main/CHANGELOG.md"

[tool.poetry.scripts]
vr180-convert = "vr180_convert.cli:app"
v1c = "vr180_convert.cli:app"

[tool.poetry.dependencies]
python = "^3.9"
rich = ">=10"
typer = "^0.12.0"
opencv-python = "^4.9.0.80"
attrs = "^23.2.0"
numpy-quaternion = "^2023.0.3"
strenum = "^0.4.15"
python-xmp-toolkit = {version="^2.0.2", optional=true}
seaborn = "^0.13.2"
ivy = "^1.0.0.4"

[tool.poetry.extras]
xmp = ["python-xmp-toolkit"]
all = ["python-xmp-toolkit"]

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
myst-parser = { version = ">=0.16", python = ">=3.11", extras = ["linkify"] }
sphinx = { version = ">=4.0", python = ">=3.11"}
furo = { version = ">=2023.5.20", python = ">=3.11"}
sphinx-autobuild = { version = ">=2024.0.0", python = ">=3.11"}
sphinx-copybutton = "^0.5.2"
sphinx-notfound-page = "^1.0.2"
sphinx-last-updated-by-git = "^0.3.7"

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
version_variables = [
    "src/vr180_convert/__init__.py:__version__",
    "docs/conf.py:release",
]
build_command = "pip install poetry && poetry build"

[tool.semantic_release.changelog]
exclude_commit_patterns = [
    "chore*",
    "ci*",
]

[tool.semantic_release.changelog.environment]
keep_trailing_newline = true

[tool.semantic_release.branches.main]
match = "main"

[tool.semantic_release.branches.noop]
match = "(?!main$)"
prerelease = true

[tool.pytest.ini_options]
addopts = "-v -Wdefault --cov=vr180_convert --cov-report=term-missing:skip-covered"
pythonpath = ["src"]

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "@overload",
    "if TYPE_CHECKING",
    "raise NotImplementedError",
    'if __name__ == "__main__":',
]

[tool.ruff]
target-version = "py38"
line-length = 110
unsafe-fixes = true
ignore = [
    "D105", # Missing docstring in magic method
    "D203", # 1 blank line required before class docstring
    "D212", # Multi-line docstring summary should start at the first line
    "D100", # Missing docstring in public module
    "D104", # Missing docstring in public package
    "D107", # Missing docstring in `__init__`
    "D401", # First line of docstring should be in imperative mood
    "D101",
    "D102",
    "D417",
]
select = [
    "B",   # flake8-bugbear
    "D",   # flake8-docstrings
    "C4",  # flake8-comprehensions
    "S",   # flake8-bandit
    "F",   # pyflake
    "E",   # pycodestyle
    "W",   # pycodestyle
    "UP",  # pyupgrade
    "I",   # isort
    "RUF", # ruff specific
]

[tool.ruff.per-file-ignores]
"tests/**/*" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "S101",
]
"setup.py" = ["D100"]
"conftest.py" = ["D100"]
"docs/conf.py" = ["D100"]

[tool.ruff.isort]
known-first-party = ["vr180_convert", "tests"]

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src/"
no_implicit_optional = true
show_error_codes = true
warn_unreachable = true
warn_unused_ignores = true
exclude = [
    'docs/.*',
    'setup.py',
]

[[tool.mypy.overrides]]
module = "tests.*"
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "docs.*"
ignore_errors = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
